kernel: Force disable HOTPLUG_CPU
CPU hotplugging is really buggy on older RT kernels, unfortunately. A patch
from newer kernels [1] mentions some of the relevant breakage, and it is
just one of several patches needed to fix the various issues. Backporting
the fixes is wholly infeasible, so just kill CPU hotplugging on RT. This
includes killing off the deep suspend state, leaving us with s2idle, which
is now used by default and cannot be changed by userspace.

As a side-effect, this also disables CAF's CPU isolation feature which
turns out to be undesirable for use on RT anyway, even after fixing it up
to work on RT. The process of isolating a CPU creates a massive jitter
spike since it keeps preemption disabled for a long time, not to mention it
doesn't save any power according to some quick measurements. It also does
not prevent isolated CPUs from getting more unpinned timers queued onto
them, so its value is dubious even on non-RT kernels.

[1] 6d337eab041d ("sched: Fix migrate_disable() vs set_cpus_allowed_ptr()")

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Faris <90097027+rsuntk@users.noreply.github.com>